Add support for callback in the WallpaperManager class
In this commit, we are adding support for a callback in the
WallpaperManager, so everytime the wallpaper is changed WallpaperManager
calls the callback passing as an argument the response returned by the
API.

We are also refactoring the WallpaperManager#setWallpaper function to
use async/await instead of the chain of thens.

Exceptions are now handled by WallpaperManager#setWallpaper, so if the
API fails the app doesn crashes.